fix: explicit type casts for nullable<uint16_t>/int16_t min/max assignments

+++ components/matter_bridge/matter_bridge.cpp
@@ -136,8 +136,8 @@ static esp_err_t create_sensor_endpoint(registry_entry_t *entry,
         humidity_sensor::config_t cfg;
         cfg.relative_humidity_measurement.measured_value =
             (uint16_t)(initial_value * 100.0f);  // Matter: hundredths of %
-        cfg.relative_humidity_measurement.min_measured_value = 0;
-        cfg.relative_humidity_measurement.max_measured_value = 10000;
+        cfg.relative_humidity_measurement.min_measured_value = (uint16_t)0;
+        cfg.relative_humidity_measurement.max_measured_value = (uint16_t)10000;
         ep = humidity_sensor::create(s_node, &cfg,
                                      ENDPOINT_FLAG_BRIDGE, s_aggregator);
         break;
@@ -147,8 +147,8 @@ static esp_err_t create_sensor_endpoint(registry_entry_t *entry,
         // Matter: Pressure Measurement in units of 0.1 kPa; hPa * 1 = 0.1 kPa units
         pressure_sensor::config_t cfg;
         cfg.pressure_measurement.measured_value    = (int16_t)(initial_value);
-        cfg.pressure_measurement.min_measured_value = 0;
-        cfg.pressure_measurement.max_measured_value = 12000;  // 1200 hPa
+        cfg.pressure_measurement.min_measured_value = (int16_t)0;
+        cfg.pressure_measurement.max_measured_value = (int16_t)12000;  // 1200 hPa
         ep = pressure_sensor::create(s_node, &cfg,
                                      ENDPOINT_FLAG_BRIDGE, s_aggregator);
         break;
@@ -160,8 +160,8 @@ static esp_err_t create_sensor_endpoint(registry_entry_t *entry,
         float lux = initial_value > 0 ? initial_value : 1.0f;
         cfg.illuminance_measurement.measured_value =
             (uint16_t)(10000.0f * log10f(lux) + 1.0f);
-        cfg.illuminance_measurement.min_measured_value = 1;      // log10(1 lux)+1 = 1
-        cfg.illuminance_measurement.max_measured_value = 65533;  // ~3.5 million lux
+        cfg.illuminance_measurement.min_measured_value = (uint16_t)1;
+        cfg.illuminance_measurement.max_measured_value = (uint16_t)65533;
         ep = light_sensor::create(s_node, &cfg,
                                    ENDPOINT_FLAG_BRIDGE, s_aggregator);
         break;
@@ -178,8 +178,8 @@ static esp_err_t create_sensor_endpoint(registry_entry_t *entry,
     case SENSOR_BATTERY: {
         flow_sensor::config_t cfg;
         cfg.flow_measurement.measured_value     = (uint16_t)(initial_value * 10.0f);
-        cfg.flow_measurement.min_measured_value = 0;
-        cfg.flow_measurement.max_measured_value = 65534;
+        cfg.flow_measurement.min_measured_value = (uint16_t)0;
+        cfg.flow_measurement.max_measured_value = (uint16_t)65534;
         ep = flow_sensor::create(s_node, &cfg,
                                   ENDPOINT_FLAG_BRIDGE, s_aggregator);
         break;
